FWWinのfullscreenモード対応にバグがいっぱいあったので，修正． src/tests/FWAppGLUTTestに，キー w を追加して， Full<->Winの切り替えに使えるようにした． あとは，warning 対応と DSTRの消し忘れの削除．
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@3469 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDConvexMeshInterpolate.cpp
@@ -57,12 +57,12 @@ Vec3f CDConvexMeshInterpolate::Support(const Vec3f& v) const {
 	struct Near{
 		int id;
 		float dist;
-		Near(int i, double d):id(i), dist(d){
+		Near(int i, float d):id(i), dist(d){
 		}
 	};
 	typedef std::vector<Near> Nears;
 	Nears nears;
-	const float minDist = 0.001;
+	const float minDist = 0.001f;
 	float sum = minDist;
 	nears.push_back(Near(curPos, minDist));
 	for (unsigned i=0; i<cn.size(); ++i){
